requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
bidict==0.23.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
fastapi-socketio==0.0.10
greenlet==3.1.1
h11==0.14.0
idna==3.10
kafka-python==2.0.2
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-engineio==4.10.1
python-socketio==5.11.4
redis==5.2.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
websockets==13.1
wsproto==1.2.0
